Problem33
Find the longest half-path between the vertices of the tree for which
the sum of the keys of the extreme vertices is minimal. Delete (by right
deletion) the middle vertex of this half-path. If there is no average
vertex (i.e., the number of vertices of the required half-path is even),
then nothing needs to be removed from the tree.

5 attempts
9/19 test cases passed