Given a binary tree.
It is necessary to check whether it is a search tree.
We will assume that there can be vertices
in a binary search tree with the same keys.
Then, in order for the tree to be a search tree,
the following requirement must be met:
for each vertex x,
all keys in the left subtree of vertex x are less than the key of vertex
x,
and all keys in the right subtree are greater than or equal to the key
of vertex x.