Problem00
A binary search tree is given in which the vertex keys are integers in the range from −2^31 to 2^31 − 1 inclusive. Find the sum of the keys of all nodes in the tree.

Problem01
Given a set of keys, build a binary search tree and perform a direct left traversal of it.

Problem02
Given a set of keys, build a binary search tree. Delete the key from it (by right deletion) if it exists in the tree. Perform a forward left traversal of the resulting tree.